+++ packages/suite/src/utils/suite/analytics.ts
@@ -14,6 +14,7 @@ import {
 } from '@suite-utils/env';
 import { getCustomBackends } from '@suite-utils/backend';
 import { AccountTransactionBaseAnchor } from '@suite-constants/anchors';
+import { BITCOIN_UNIT_ABBREVIATION } from '@wallet-hooks/useBitcoinAmountUnit';
 
 import type { AnalyticsState } from '@suite-reducers/analyticsReducer';
 import type { UpdateInfo } from '@trezor/suite-desktop-api';
@@ -38,6 +39,7 @@ export const getSuiteReadyPayload = (state: AppState) => ({
     enabledNetworks: state.wallet.settings.enabledNetworks,
     customBackends: getCustomBackends(state.wallet.blockchain).map(({ coin }) => coin),
     localCurrency: state.wallet.settings.localCurrency,
+    bitcoinUnit: BITCOIN_UNIT_ABBREVIATION[state.wallet.settings.bitcoinAmountUnit],
     discreetMode: state.wallet.settings.discreetMode,
     screenWidth: getScreenWidth(),
     screenHeight: getScreenHeight(),