+++ packages/suite/src/views/settings/general/BitcoinAmountUnit.tsx
@@ -1,18 +1,30 @@
 import React from 'react';
-import { ActionColumn, ActionSelect, SectionItem, TextColumn } from '@suite-components/Settings';
+import { analytics, EventType } from '@trezor/suite-analytics';
 import { PROTO } from '@trezor/connect';
+
+import { ActionColumn, ActionSelect, SectionItem, TextColumn } from '@suite-components/Settings';
 import { Translation } from '@suite-components/Translation';
 import { useAnchor } from '@suite-hooks/useAnchor';
 import { SettingsAnchor } from '@suite-constants/anchors';
-import { useBitcoinAmountUnit } from '@wallet-hooks/useBitcoinAmountUnit';
+import {
+    BITCOIN_UNIT_ABBREVIATION,
+    useBitcoinAmountUnit,
+} from '@wallet-hooks/useBitcoinAmountUnit';
 
 export const BitcoinAmountUnit = () => {
     const { bitcoinAmountUnit, setBitcoinAmountUnits, UNIT_LABELS, UNIT_OPTIONS } =
         useBitcoinAmountUnit();
     const { anchorRef, shouldHighlight } = useAnchor(SettingsAnchor.BitcoinAmountUnit);
 
-    const handleUnitsChange = ({ value }: { value: PROTO.AmountUnit }) =>
+    const handleUnitsChange = ({ value }: { value: PROTO.AmountUnit }) => {
         setBitcoinAmountUnits(value);
+        analytics.report({
+            type: EventType.SettingsGeneralChangeBitcoinUnit,
+            payload: {
+                unit: BITCOIN_UNIT_ABBREVIATION[value],
+            },
+        });
+    };
 
     return (
         <SectionItem