+++ packages/suite-analytics/CHANGELOG.md
@@ -8,10 +8,13 @@ Added:
 
 -   suite-ready
     -   labeling: 'dropbox' | 'google' | 'fileSystem' | 'sdCard' | ''
+    -   bitcoinUnit: 'BTC' | 'μBTC' | 'mBTC' | 'sat'
 -   settings/general/labeling
     -   value: boolean
 -   settings/general/labeling-provider
     -   provider: 'dropbox' | 'google' | 'fileSystem' | 'sdCard' | ''
+-   settings/general/change-bitcoin-unit
+    -   unit: 'BTC' | 'μBTC' | 'mBTC' | 'sat'
 
 ### 1.20